Chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.134.0",
    "google-genai>=1.65.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
